test: move 'jupiter' package in 'reactive'
